Bind caches as eager singletons
All H2 caches must be created before the enclosing LifecycleManager
starts, therefore they must be created eagerly. If they are not, it is
possible for an assisted factory to attempt to create the cache for
the first time only when that factory's create() method is called (as
opposed to when the factory is injected).

Change-Id: I058d8783161ffc4274dd78337842ce55c53de183

Former-commit-id: 0f62e836d57d3032b6b0c115ad3297e7eec6ffd3