Remove direct Lucene dependency from InMemoryModule
Custom secondary index implementations should be able to use this
module for writing index tests without linking against Lucene.
However, for other potential uses of this module for tests within
Gerrit, it makes sense to still use Lucene by default without
installing any additional modules. So, as a hack, keep the Lucene
dependency but construct it via reflection.

The same argument could be made about the H2 cache dependency, but the
specific use case we are interested in at the moment is the index
implementation, so extract that one first.

Change-Id: I7d213a9f233b819ec0bc6632323e65ac9b270200

Former-commit-id: a1fe8214703f9cf38a6adf93b0013a41f9199ed3