SideBySide2: Synchronize horizontal scrolling
Disable CM3 line wrapping and rely on synchronized horizontal
scrolling instead.  This fixes alignment problems caused by long lines
that exist only one of two sides.  When line wrapping, the view did
not pad the other side to account for the extra line height used by
the wrapped line, skewing everything after it.

Since horizontal scrolling is now functional change the widths of the
views to be related to the width of the browser window.  Keep a
maximum width to prevent the a large gap between the sides on very
wide monitors.

Change-Id: I3a3b10fefdacd7246e4856365cb91d934f0346e2

Former-commit-id: 5bf4addf76d8fe7bff67401116acd03783bbd33f