Use TimeUnits for sort key calculations
The unspecified units were confusing when writing tests around these
values, so be explicit about units in variable names and quantities
for conversion.

Use Ints.checkedCast so we fail fast when inserting a change 4000+
years in the future rather than having it not show up in search
results.

Correct the sort key epoch time, which likely had a timezone error
when initially calculated:

$ python -c 'import datetime; print datetime.datetime.fromtimestamp(1222819200L)'
2008-09-30 17:00:00

This mistake does not affect any data; the comment in the code was
just confusingly incorrect.

Change-Id: I6d8bcf4985e91c3d115002db42642049da8c708f
Former-commit-id: a2023e948652592647bdd14909e500e6acaac6c8