Use DateTimeUtils.currentTimeMillis() to get current time
This method from joda-time supports replacing the system time provider
with a custom provider for tests. Since it is verbose, and we expect
more related methods, put it in the more succinct TimeUtil.nowMs().

This commit covers the trivial cases in the server side; client-side
code still uses System.currentTimeMillis().

Change-Id: I6c56e8c5bbb0cf7b0271e431d1ebdb532967b9e8

Former-commit-id: 46b1ac8f070f7808afb86d45db7c98ac819745b9