Increase accounts listed on Become Any Account login screen to 100
The number of accounts listed on the Become Any Account login page is
currently hard-coded to 5.

This can be inconvienent for developers on systems that have a larger
number of test accounts.  Of course the developer can manually enter
the account name, but it would be nicer if the screen could show more
and allow the developer to just click.

Increase the number of accounts shown to 100.  This should be enough
for most test systems.

Change-Id: I3a17c17e18b2226ab4986306fc3cbf071cadd322

Former-commit-id: c8fb8ed522604db092806528de6ba44db49a3335