ChangeScreen2: Add query icon for project
As on the old change screen the project link should link to the
project administration page and there should be a query icon to search
for changes on the project. Clicking on the query icon opens the
default dashboard of the project or if no default dashboard is defined
it queries for the open changes of the project.

This behaviour is consistent with the old change screen so that users
don't get confused.

Bug: issue 2075
Change-Id: I62f87466b60896ecb046e6053f8361405a5c4285
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: a2b2a2703ec8e200571cd2c8978e26345a18804e