Set git submodules back to relative paths
In 12bc68fce248587a9d4aef138d81f87b331f6c1f [formerly 246f9b52854fc1b11dbe14085e2c0ac0f6ddda95] the submodules were
changed to use absolute paths, to prevent problems on mirrors
where only the gerrit project is present, or where the plugin gits
are not on the same relative path as they are on gerrit-review.

This however causes problems for mirrors where the plugins *are*
present and in the expected relative path:  the plugins are still
cloned off the gerrit-review server, rather than the mirror.

Switch them back to relative paths to fix this.

Change-Id: I30b5e6f96fbbc538665be9a618c3cf092e1588f1

Former-commit-id: 6d093932ce43e4bd47b12d09426f3479cbcff7b1