Support project specific plugin configuration in own config file
For some plugins the project specific configuration is too complex to
be stored in a single 'plugin' subsection in 'project.config'. If a
plugin needs subsections in its project specific configuration then
the project specific plugin configuration can be stored in an own
configuration file in the project's 'refs/meta/config' branch.

This change adds a method to PluginConfigFactory to easily access the
project specific plugin configuration that is stored in such a plugin
configuration file.

The project specific plugin configuration is cached within
ProjectState (which is cached in the project cache). This avoids
reloading and reparsing of the config file on every access. After a
project is evicted from the cache the project specific plugin
configuration is reloaded. This means that any update that is done to
the plugins configuration file is immediately active since any push to
the `refs/meta/config` branch evicts the project from the cache and
hence triggers a reloading of the project specific plugin
configuration.

Change-Id: Id5d371a22041bd7381f92e75fb021af5318249b4
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 705f2848636e13849157dc2d0e1b5add38990134