Copy reviewed flag for deleted files
If a file is deleted and a user marks it as reviewed, copy the
reviewed mark onto newer revisions provided the same content has
been deleted. Checking the content of the deleted file allows
reviewers to re-examine a delete after a rebase, in case there is
content in the newer deleted revision that needs to be considered
to be moved elsewhere.

Rewrite the copying process to iterate the 2 or 4 trees exactly once,
relying on a PathFilterGroup to identify the relevant files. This
should be faster on larger changes as it avoids revisting trees. It
is also easier to handle the 2 additional trees for the ancestor
content check on deleted files.

Change-Id: Ic710ca2971e66ea919ee1c9f51eb42538174e7b3
Former-commit-id: a7ef8dadbc01f77dc088cbd5aded417f0d4c5c44