Fix inclusion of LF in IntralineLoader
An edit may contain consecutive LFs. Currently, if a replacement
region ends with the first LF, the second LF may still be included,
producing a region that spans two lines. If there is another region that
starts with this second LF, the end of the first region will be
greater than the beginning of the second one, which is an error.

The error does not seem to affect the old diff view but is caught
by SideBySide2. Fixed by checking if the region already ends with LF.

Change-Id: I86aa9100dc1318935aa350a0aba0c411e2dc5841
Former-commit-id: f967024c3c63af871b919441f07161a20092f44a