Show in change table and user dashboard if change is not mergeable
For reviewers it's an important information if a change is mergeable
since eventually it doesn't make sense to review a change that cannot
be merged. This is why this information should be immediately visible
in the change table and in the user dashboard.

As a contributor I want to see on one look which of my open changes I
have to rebase since they are not mergeable anymore. It helps me a lot
to see this information in my user dashboard.

Change-Id: I881452ca0f1905b7d4f0c026f98a42ac4d5f58bd
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 56f39eb0a0e5cfd33276a7889faf09ed191ff035