canReadCommit: further improve performance by using IncludedInResolver
When searching for a commit the IncludedInResolver avoids traversing the
same commit (sub)graph for the second time if it already failed to find
that commit once in that (sub)graph.

Partition the refs into two sets:

  after  = commits with time >= target.time
  before = commits with time <  target.time

It is highly likely that the target is reachable from the "after" set.
It could be reachable from the "before" set only in case of a clock skew.
Sorting both "after" and "before" sets by time will increase the
probability that a cached failure to find a commit will be reused when
traversing the commit graph from newer commits.

Before invoking the IncludedInResolver, filter out all refs for which the
user doesn't have the READ access. Also, avoid fetching all refs from
the ref database.

This change required some refactoring of the IncludedInResolver class.

Change-Id: Ic5651bb0b71701ad64920e195ab735038edef377

Former-commit-id: b1944c0fc0637fcab03e636a40092b2f252980ab