ChangeScreen2: Show inline comments in history
When a history comment is expanded include the inline comments written
on that revision by the author at the same timestamp.  This makes it
easier to follow a change's evolution over time, especially if the
reader is new to the change and does not have the entire email thread
in their inbox.

Bug: issue 93
Change-Id: I57fe29fea2b3651306ff0a58746cefcfe1669267

Former-commit-id: eb924ff74a90bbeb37cd3856dae4e9325598656b