Default to CM3 based SideBySide2 if using ChangeScreen2
This allows hyperlinks in emails (which lack the ,cm suffix) to open
the CM3 based UI if the user or host has configured ChangeScreen2
as the preferred/default UI version.

Add ",sidebyside" as an alias for the old version of the UI as an
emergency escape hatch in case a user needs to get back to the old
file UI temporarily.

Change-Id: I3dd3a893429020b56d641a42b2738179b9416c85

Former-commit-id: ce8620d7b3cb3a9d3c724f4837acd8b1e405a20e