SideBySide2: Fix minor key help layout issues
Move <Enter> to be with 'o' in the Actions section.  These keys do the
same thing and they should not be in different key groups.

Register Actions after Comment Editing.  The sections alternate
left/right when packed into the dialog so this puts Actions below
Navigation for signed in users that also have Comment Editing.

  Application       Jumping

  Navigation

  Actions           Comment Editing

The overall flow reads better with this ordering.

Change-Id: Ib68f92753c81732e99ae31e6cf5a5f37190d4762

Former-commit-id: 7df9f202426a45189fe07237d7f0addf8e5e950b