SideBySide2: remove unnecessary padding
There was a bit too much space at the bottom of the window when
SideBySide2 is open and the headers are "scrolled off".  Its caused
by using 10px in the padding computation.  5px is sufficient to keep
everything in view. Use 5px, giving 5px back to CM3 and the rest of
the UI.

Change-Id: Ic77c4f63d014d4d813bf0d81c4b7e963d11ae6c1
Former-commit-id: d55eef420bd8073e5de4ade000fc49b07e164543