ChangeScreen2: Fix related changes tab from expanding too tall
The height of the tab bar header must be subtracted from the maximum
height given by the other columns in the header, otherwise a related
changes tab can cause the UI to shift the "Files" table down when
it is selected.

Remove the padding on the bottom panel, it saves some space and
ensures the header substraction is accurate.

Change-Id: Ie2ce21bb81c1ca56c1fe13af433d76e5d8508b16

Former-commit-id: c4a3e5f83806e5207924e5436250fa58e0f78595