Use rev-parse to find gitdir when generating commit-msg hook hint
This updates the Change-Id validator hint to advise the user to use
"git rev-parse --git-dir" to find the proper gitdir, in case the
user is using submodules.  Also, the gitdir is stored into a temp
variable, because msysgit has an unusual behavior related to the
path returned by rev-parse otherwise.

Change-Id: Ie4fcd0a73bd1cee12a052f38541f83aa8821c886
Former-commit-id: dd57e969b7bba785a90369f996b89f30d2e406f0