ChangeScreen2: Add padding to collapsed history
2px above/below each collapsed summary line makes it much easier to
read the text, as the lines blur together less easily to the reader.

With an additional 4px in height available increase the avatar in the
collapsed state to be 20px instead of 16px.  This makes it a bit
easier to resolve some avatar images and identify users.  Position the
avatar over the border in the collapsed state to prevent artifacting
the rounded edge.

Change-Id: I5ba9820a5b084f9193d6faa9b4088d0080d19fa7

Former-commit-id: f61e5a07d45931ec5a26f6d72189bedc3b947dc5