ChangeScreen2: Move star to far right of status bar
Web applications tend to place the "star action" for an object on
either the far left or the far right. Placing it in the center makes
it hard to locate.

Move the star onto the far right of the status bar.  Placing it on the
far left offsets the commit message subject from the commit message in
an ugly way, so far left position is not an option.

Change-Id: Ia477fbfe08fae7b08065d2c2918eb34558f1d89a

Former-commit-id: bc7084f660b515fce0dd6a9609b9ffa81a10140b