SideBySide2: Fix misplaced widgets at bottom of window
In 89f0280622 [formerly 5b83d94271] ("SideBySide2: Batch updates to CodeMirror3 during
onShowView") Colby and I tried to improve rendering performance by
avoiding multiple redisplays of the CM3 widgets during onShowView().

That change worked to reduce DOM update cost, but the batched
setHeight and setOption caused the widgets to be drawn at the
bottom of the window when the file first opened.

A LineWidget needs to be added to diffTable to respond to browser
events, but CodeMirror doesn't render the widget until the containing
line is scrolled into viewportMargin, causing it to appear at the
bottom of the DOM upon loading. Fix by hiding the widget until it is
first scrolled into view (when CodeMirror fires a "redraw" event on
the widget).

With this fix, viewportMargin does not need to be set to infinity
anymore. Leave it at the default of 10 lines. During onShowView only
the height has to be updated, so wrapping the update in an operation
is no longer required.

This fixes the skip bar and comment widget display bugs, apparently
without reintroducing the latency Colby observed on Thursday.

Change-Id: Id55232853bcfde12d065fd88b69f3538115f2dfe

Former-commit-id: e07d21e7fb208cebb057b5dbb2abc81eb86cb9fa