ChangeScreen2: Handle missing current revision
If for a change the current revision is not returned ChangeScreen2
fails to load and the "Working ..." label is displayed forever.

If the current revision is missing, diplay the last revision that was
returned. If no revision was returned display an error.

The current revision is missing for changes where the current revision
is a draft patch set that is not visible to the calling user.

There are also corrupt changes that return no revision at all.

Change-Id: I9a063b4dbc6ad7ae5b67d9a022b2141ce66ead32
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 693fdd253f75f37106877e7f95182c619d85463d