Ensure line breaks are kept when replying to a multi line comment
Add a leading space to each quoted line to ensure that line breaks are
kept.

If the existing line breaks are kept it can happen that some lines get
very long (e.g. if a whole paragraph was typed without an explicit
line break). To avoid this long lines are automatically wrapped.

Change-Id: I397e10f68908d85049abeb0e64cc923736f060bf
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: fe073f3bd7513542fcb3b2637f000bd6eee163a3