Visualize sizes of changes in change tables and user dashboards
Add a new column to change tables and user dashboards in which the
size of a change is visualized by a colored bar. For small changes a
short green bar is shown, for medium sized changes a middle sized
yellow bar is shown and large changes have a full red bar.

Visualizing the change sizes gives a hint to the reviewers about how
complex a change is and how long it will take to review it.

For contributors the visualization of the change size may be a
motivation to split large changes into smaller pieces which will speed
up the review process.

How many changed lines are considered as a large change is
configurable.

Change-Id: I88497b8aea00047d93ad7d079eecc5e1052187d2
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 662e55f6af114c7530e08abce3d9bfe46d3d9882