SideBySide2: Save all drafts when leaving / semi-auto focus on drafts
Keep track of all unsaved drafts and save them when leaving the page.

Using CodeMirror's "redraw" event, autofocus on the editArea of a
Draftbox when scrolled into viewportMargin. This works smoothly for
the common case where a user starts typing a draft, scrolls around to
view the file, and scrolls back to continue typing. Even if the user
forgets to scroll back and accidentally pressed "u" or "a", the draft
will be saved, minimizing potential frustration.

Change-Id: Idde88aaeb01e46f85734b48caad1de7c9b2ffba3

Former-commit-id: 44adfee873378af6d322b47121bdcfcc1423dc6f