Update mergeability flags if submit type or submit rule is changed
The mergeability flag for open changes is now recomputed if the
project submit type, the project content merge setting or the project
submit rule is updated. In this case the recomputation of the
mergeability flag is enforced because the stale computation in the
Mergeable REST endpoint is only based on the SHA1 from the destination
branch, but we don't cache for which submit type, content merge
setting or submit rule the mergeability flag was computed.

Change-Id: I525c542d17674a974db456f98f801822992c7bb8
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: 76ba20b2a72abd5ede06bc2fff8502ff69c8f223