Enable plugins on init to read/edit the All-Projects project.config
Some plugins may store their configuration in the project.config file
of the All-Projects project. The configuration may depend on user
input during the plugin initalization. Plugins are now able to write
configuration parameters into the All-Projects project.config during
an InitStep.

E.g. plugins may provide functionality which should be
enabled/disabled per project, which is reflected by an 'enabled'
parameter in project.config:

  [plugin "myPlugin"]
    enabled = true

If this parameter is not set, the value should be inherited from the
parent project. Having the possibility to edit the All-Projects
project.config on init the user can decide during the plugin
initialization if the functionality should by default be
enabled/disabled for all projects.

Change-Id: I8e240858e0778469c0d511ff380865f317448e15
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: d4cfac1cd668c83ce0ba53d4ca89faf78a048beb