Fix getting inherited values from PluginConfig
The code for constructing a Config with the parameters from the
current project plus the missing parameters from the parent projects
relies on getting all parameter names from Config#getNames(...). The
problem is that Config#getNames(...) doesn't include parameter names
which are only available in the default Config (the Config that we
provided to the constructor when creating a new Config). This may lead
to a situation where a parameter is defined on a project, but a
grandchild of it is not inheriting this parameter although it and its
parent do not have this parameter. To fix this we are now copying the
initial Config instead of creating a new Config and providing the
existing Config as default Config to it.

Change-Id: I93ed8f7f13f1bc757e59a5897f1949540c41f9ed

Former-commit-id: a7eaf144017d4e28cf08eb9b7d15a06ac80af8e4