Fix Daemon start when sshd.listenAddress=off and host key doesn't exist
When a site is initialized with sshd.listenAddress=off, a ssh host key
is not generated. The Daemon startup failed in that case because it
tried to start the SshDaemon and the injection of a KeyPairProvider
failed because the ssh host key didn't exist.

Set the Daemon.sshd option to false also when sshd.listenAddress=off
to ensure the same startup logic as when --disable-sshd option is used.

Change-Id: Ia9a64f77ca4dc977aafa061a9b9a736db947d5df

Former-commit-id: 91205ae8761fddeb7d0a854c572064cef6476bd4