Add Gerrit.on('submitchange', f) to JS plugin API
Allow JavaScript plugins to register additional handlers onto the
submit button on the change screen, and reject the operation if the
plugin does not think it is reasonable to invoke submit from the
interactive web UI at this time.

chromium-review uses a chunk of JS to warn developers if the tree
is blocked due to a build error.  Fixes for the build need to be
submitted anyway, while most other changes should be avoided:

  Gerrit.on('submitchange', function (c, r) {
    return confirm('Tree is blocked, submit anyway?');
  });

Change-Id: I5a7fc94654ce47f83d250a31d5384d46744a2954

Former-commit-id: 7c418993fc81666ba8b7f4962b44679c41ecb0e0