Add Gerrit.on('showchange', f) to JS plugin API
chromium-review.googlesource.com uses a chunk of JavaScript to monitor
the anchor, detect if the URL is #/c/...  style, and then extracts the
project name out of a table cell.  This is a horrible way to add web
UI logic based on the project.

Add an event to ChangeScreen2 that plugins can register against.  Pass
the ChangeInfo so functions can use native properties like "project"
instead of scraping the displayed web interface.

Change-Id: Ie6daed2598d4e4f9c9fe0dee40ab2e36b81012bf

Former-commit-id: db284cd1526eef2e190926169370aab8c5a86542