Build Main for Java 1.2
This tiny class uses a very ancient and limited API so it runs on any
JVM. Its entire purpose is to check the minimum Java version, fail if
it is too low, otherwise jump into the real class in another package.

The Maven build used to create this class for Java 1.2, so it would be
helpful to users attempting to launch Gerrit on the wrong JVM. That flag
was lost when the build moved to Buck.

Reset the source and target level to 1.2 so the warning message might
be useful again, especially during the jump from Java 6 to Java 7.

Change-Id: I9187e13ee67c9c482a8e11296649fcf712e5aad1

Former-commit-id: ab21a266b19bf6d02e13d7a8987b000f8d1d7233