Add REST endpoint to get content of a file from HEAD of a branch
By GET on /projects/<project-name>/branches/<branch>/files/<file>/content
it is now possible to retrieve the content of a file from the HEAD
revision of a certain branch. The file content is returned as base64
encoded string.

This endpoint is useful for integrating tools that need access to
certain files, e.g. a tool may need to get the pom.xml from every
project. In this case it would be too much overhead to clone each
repository just to get this one file from it.

Change-Id: I70f294fbcc8a4cc88cd3c1265653cc9870b6433e
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>

Former-commit-id: d31e5586ad5e7b936ba7535e192c05653d82c7fe