Hyperlink to groups in access editor
Internal group names are automatically hyperlinked in the access
control editor. This makes it easy to view the members of the group
and determine if access is correctly assigned.

Plugins supplying group systems may include a url to a group
information page. If present use these URLs as hyperlinks in the
access editor, providing the same functionality.

Change-Id: I9605fdc1d43d6705a01ec2b68b083e8678dbc88a

Former-commit-id: 0a70c69cb82b89f423c126fc203db9cbbf4243a6